inital prototype of the "extended" hardware information
Especially when testing it seems quite important to be able to gather some information about the hardware like temperature, memory usage, cpu usage etc.

The hard part is that all of those things depend on the os/hardware etc, so it's generally a mess. Main focus lies on Debian (and Raspian) and Windows.