Disallow Js in tsconfig.json
Having both the typescript and javascript files can confuse the TS-MSBuild (producings erros). By disallowing "normal" JavaScript this error gets fixed, in addition to the enforced TypeScript usage.

In the long run the webapp should probably not live in the wwwroot of the webserver project, but I haven't yet decided on the details.