Disable CS1591 and SA1600 (Enforcing comments)
Not sure if this is sensible, but I still feel like doing it anyway.
I don't need a comment on the Name propety to know it's a Name, it just reduces the Signal-to-Noise-Ratio. If I can't tell it from the Name alone, it's probably a bad name.

At some point I will have to do something about Documentation for the HTTP-API, but I don't think the current xml Comments are really fit for that purpose.