Process images for an engine in parallel
Also fixes a small bug - iterating over the engines without ToList() kept a lock on the engine table. If anything tries to write to the table (like acquiring an engine does) it ended in a deadlock (until the timeout kicks in).

The ImageOptions are ordered in what I expect to take the longest to the shortest in the hope that the long-running ones get processed first.
It now takes a bit more than half the time. 4c/4t to 12c/24t didn't appear to make a large change.  Processing engines in parallel might yield more speedups on many core systems, albeit also requiring more complicated synchronisation.