Upgrade ToolkitMvvm, reenable SA1633
ToolkitMvvm seems to have changed from a microsoft thing to a more explicit community thing, so it also got a new NuGet-Package with the new name.

The Bug related to SA1633 is only fixed in 8.0, so I'm using the preview version for now.

This commit also removes project-specific no-warns and relies on the ruleset instead.
The ReSharper/Rider config is updated to ignore CS1591 too.

Resolves #181