Implement forget loco command for DccExSerialSystem
Resolves #235

Dcc++ EX sends a estop when forgetting to the engine being forgotten (actually twice). So I would expect engines to actually be stopped when forgotten about.

When a engine is released that still has a non-zero speed the "forget loco" command is not sent - under the assumption that a estop isn't desired and a user will take over control again soonish.