Remove the DefaultConfiguration and move them into the Option class
Doing it this way is slightly more straighforward and doesn't rely on the string keys being correct in all the places the configuration is used.

The biggest advantage is that the Options class can be annotated for nullability. With .NET 7, all config values are nullable (which is the correct thing to do), the DefaultConfiguration inMemory config source simply cannot be used for the nullability analysis.

Using options for the CommandStation parts needs some more thinking.