TestDbTool - add two random tags to every engine
The two tags simulate using the tags as "alternative names". I personally add the full UIC-number in there - that's where the idea comes from.

This really should really inflate the table for tags - from my own testing it doubled the time a query takes on a 10K db to around 40ms.